Pullup ticket #2804 - by he mk/bulk/build: robustness improvement
Revisions pulled up:
- mk/bulk/build			1.106
---
Module Name:	pkgsrc
Committed By:	he
Date:		Thu Jul  9 12:46:50 UTC 2009

Modified Files:
	pkgsrc/mk/bulk: build

Log Message:
With the most recent pkg_install, directories are removed when the
last file in a directory is removed.  This might cause common and
expected-to-be-existing directories in /usr/pkg to be removed, and
a subsequent attempt at installing a file to the now non-existent
directory will instead create a file with the name of the expected
directory.

This will create PLIST errors for the package in question, but also will
cause the erroneously named file to not be removed on package removal.

This can cause cascading bulk build failures for subsequent packages.
To prevent this, after each package is done, check if some of the
common top-level expected-to-be directories are now files, and
remove them and emit an error message if so.

I *think* I already fixed the single package which had the unfortunate
problem of not declaring "include" as an installation directory, but
this should prevent the problem from re-occurring in the future as well.

Thanks to joerg@ for the hint for pulling in the value of PREFIX.